requirements.txt
asgiref==3.4.1
Django==3.2.7
django-discover-runner==1.0
django-filter==2.4.0
django-role-permissions==3.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
